[#5333 1/N] Snapshot replacement models + queries
Start by separating out the DB models, queries, and schema changes
required:

1. snapshot replacement records

Similar to region replacement records, snapshot replacement records
transition through the following states:

        Requested   <--
                      |
            |         |
            v         |
                      |
        Allocating  --

            |
            v

         Running

            |
            v

     ReplacementDone  <--
                        |
            |           |
            v           |
                        |
        Completing    --

            |
            v

        Completed

which are captured in the `SnapshotReplacementState` enum. Transitioning
from Requested to Running is the responsibility of the "start" saga, and
transitioning from ReplacementDone to Completed is the responsibility of
the "finish" saga. Transitioning from Running to ReplacementDone is
performed once all Upstairs for impacted volumes are notified of the
relevant replacement.

State transitions are performed in a similar way to how region
replacement record states are transitioned.

2. snapshot replacement step records

Once a snapshot's volume has had the relevant target replaced, any still
running Upstairs that was constructed with that target requires a
replacement call to be sent. A "step" record is created for each of
these volumes, and will transition through the following states:

        Requested   <--
                      |
            |         |
            v         |
                      |
         Running    --

            |
            v

        Complete

Only when all step records for a request are in state "Completed" will a
snapshot replacement record transition to "ReplacementDone".

3. also, the `volume_replace_snapshot" function

The `volume_replace_snapshot` function is also included in this PR.