[#5333 3/6] Snapshot replacement start
This commit adds a "snapshot replacement start" background task that
will:

1) check for snapshots that are stored on disks that were expunged, and
   insert snapshot replacement requests for them.

2) check if there are any snapshot replacement requests in the
   `Requested` state and run the new "snapshot replacement start" saga
   for them. This background task will also pick up manually requested
   snapshot replacements.

Also in this commit is the "snapshot replacement start saga", which will
transition a snapshot replacement reuest from `Requested` to
`Allocating` and:

- allocate a new Region (with source set to Some so that a clone occurs)

- create a blank Volume that will be used to stash the snapshot to
  delete later

- swap the snapshot being replaced with the new region that was cloned
  from the others in the read-only region set

- update the snapshot replacement request's state to "ReplacementDone"
  and clearing the operating saga id

This represents the first step to be taken after a snapshot goes away:
allocate the replacement, and swap it in to the affected snapshot
volume. Once this is done, anything that uses the snapshot volume as a
read-only parent will no longer reference the expunged snapshot, and
will work without any issues.

Existing constructed Volumes running in a propolis or pantry context
will remain unmodified: a future commit will contain the saga that takes
care of performing the necessary read-only target replacements.