Add endpoints reading for snapshot and image data
This PR adds endpoints to Nexus that allows users to read the contents
of two sources of read-only customer data: snapshots and images.

Unlike with disk import, users will not have to manage any
implementation details. Nexus is entirely responsible for attaching a
copy of the read-only object's volume to a Pantry and proxying block
read requests.

A background task will detect when a snapshot or image needs to have
this associated attachment, and invoke the appropriate attachment saga.
Likewise the same background task detects when a snapshot or image is
deleted, and invokes the appropriate detach and delete saga. Currently
every non-deleted snapshot and image will have a volume copy attached to
a random Pantry: the memory required for this is minimal due to the
activation only being read-only, and only serving reads. Measurements
showed RAM usage in the order of ~140MiB.

Internally, "user data export" objects store the association between the
snapshot or image, the read-only volume copy that was made, and the
randomly selected Pantry where that volume copy was attached and
activated.

If a Pantry is expunged, any of user data export objects that use that
Pantry will need to be deleted and recreated using other Pantries. The
aforementioned background task will use DNS (instead of a rendevous
table, as one does not exist for the Pantry) to query for the list of
Pantries that are in-service, and use that to invoke the necessary
detach and delete saga.

Fixes #5824