[5/5] User data export: endpoints
This final commit in the series adds "read" endpoints for both snapshots
and images, fills in support in Nexus for this type of export, and adds
integration tests.

Read requests for a resource are proxied to the Pantry and volume ID
that resource's associated user data export object recorded. These new
endpoints respect the Range header in the request: if not supplied, then
the whole object will be exported, and if supplied, only a sub-range is.

Nexus will return an error if:

- the user data export object is not created yet
- the Pantry that was used for that export was expunged
- the Pantry that was used was restarted and lost all state.

The last case will cause the user data export record to be marked as
deleted - it is a case that the associated background task will not
notice and marking the record as deleted will cause the record (and
attachment) to be recreated.

Fixes #5824.