Add a user provision type to silo user
Silo users have traditionally equated "external id" with whatever the
silo's identity provider assigns for a user: username, email, etc. SCIM
has an external id attribute that is distinct from other related
attributes that could be sent, including username and email.
Importantly the SCIM client that is creating a user could send a NULL
external ID, which is not supported by the current field's database
type.

The first step to altering the silo user model to support SCIM's extra
attributes is to add the silo's user provision type to the silo user
model, and use this to change the semantics of the external id field
depending on what provision type is used.

The second step that will help with type safety is to use the new user
provision type field to separate the silo user database model into
distinct silo user types at the datastore level. Call sites now have to
use either the ApiOnly or Jit variant of the higher level silo user
type. With the addition of user provision type to the silo user model,
Nexus can also now check that the provision type on the user matches the
silo, or (like with some of the local idp functions) accept only one
specific variant.

One possible bug that this commit addresses is that the creation of the
recovery silo did not have a check to ensure that the created silo was a
"LocalOnly" identity type.  The code prior to this commit could create a
SamlJit silo, and the creation of the associated silo user would
succeed.